Updated file copy

+++ src/main/java/eu/tsystems/mms/tic/testerra/plugins/teamcity/TeamCityHistoryDownloader.java
@@ -80,7 +80,11 @@ public class TeamCityHistoryDownloader implements Loggable {
             File historyFile = this.downloadHistoryFile(historyFilePath);
             Report report = Testerra.getInjector().getInstance(Report.class);
             File finalReportDirectory = new File(report.getFinalReportDirectory(), REPORT_MODEL_DIRECTORY);
-            FileUtils.cleanDirectory(finalReportDirectory.getAbsoluteFile());
+            if (finalReportDirectory.exists()) {
+                FileUtils.cleanDirectory(finalReportDirectory.getAbsoluteFile());
+            } else {
+                Files.createFile(finalReportDirectory.toPath());
+            }
             Files.move(historyFile.toPath(), finalReportDirectory.toPath(), StandardCopyOption.REPLACE_EXISTING);
         } catch (IOException e) {
             log().warn("Cannot download history file to report directory: {}: {}", e.getClass(), e.getMessage());
